requirements.txt
Django==2.0.5
#django-blog-zinnia==0.19
git+https://github.com/Wtower/django-blog-zinnia.git
django-ckeditor==5.4.0
django-constance==2.1.0
django-debug-toolbar==1.9.1
django-imagekit==4.0.2
django-jet==1.0.7
#django-modeltranslation==0.12.1
git+https://github.com/deschler/django-modeltranslation.git
django-picklefield==1.0.0
django-rosetta==0.8.1
feedparser==5.2.1
google-api-python-client==1.4.1
mock==2.0.0
python-memcached==1.59
pytz==2018.4
Unipath==1.1
lxml==4.2.3
django-admin-rangefilter==0.3.6
celery==4.2.0
redis==2.10.6
django-import-export==1.0.1
whitenoise
cfscrape